Feat: SearchProvider(context Api) 적용
