--- src/views/Recipes/SearchRecipes/SearchViewRecipes.tsx
@@ -1,10 +1,0 @@
-import SearchCondition from './SearchCondition';
-import { S_RecipeContainer } from '../../../styles/RecipeContainer';
-
-export default function SearchViewRecipes() {
-    return (
-        <S_RecipeContainer>
-            <SearchCondition />
-        </S_RecipeContainer>
-    );
-}
